Refactor shortcuts tree out of navmenu so we can use it in the shortcuts editor
