+++ packages/web/components/templates/navMenu/LibraryMenu.tsx
@@ -22,13 +22,13 @@ import { HomeIcon } from '../../elements/icons/HomeIcon'
 import { LibraryIcon } from '../../elements/icons/LibraryIcon'
 import { HighlightsIcon } from '../../elements/icons/HighlightsIcon'
 import { CoverImage } from '../../elements/CoverImage'
-import { Shortcut } from './NavigationMenu'
 import { OutlinedLabelChip } from '../../elements/OutlinedLabelChip'
 import { NewsletterIcon } from '../../elements/icons/NewsletterIcon'
 import { Dropdown, DropdownOption } from '../../elements/DropdownElements'
 import { useRouter } from 'next/router'
 import { DiscoverIcon } from '../../elements/icons/DiscoverIcon'
 import { escapeQuotes } from '../../../utils/helper'
+import { Shortcut } from '../../../lib/networking/shortcuts/useShortcuts'
 
 export const LIBRARY_LEFT_MENU_WIDTH = '275px'
